+++ Elements/src/Wall.cs
@@ -1,5 +1,6 @@
 using Elements.Geometry;
 using Elements.Geometry.Solids;
+using Elements.Interfaces;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
@@ -10,7 +11,7 @@ namespace Elements
     /// A wall defined by a planar profile extruded to a height.
     /// </summary>
     [UserElement]
-    public class Wall : GeometricElement
+    public class Wall : GeometricElement, IHasOpenings
     {
         /// <summary>
         /// The height of the wall.
@@ -21,6 +22,11 @@ namespace Elements
         /// The profile of the wall.
         /// </summary>
         public Profile Profile { get; protected set; }
+
+        /// <summary>
+        /// A collection of openings in the wall.
+        /// </summary>
+        public List<Opening> Openings { get; } = new List<Opening>();
 
         /// <summary>
         /// Construct a wall by extruding a profile.
@@ -86,7 +92,7 @@ namespace Elements
                                            isElementDefinition,
                                            id == default(Guid) ? Guid.NewGuid() : id,
                                            name)
-        {}
+        { }
 
         /// <summary>
         /// Construct a wall from geometry.